Reorganize the execution pools under SolidQueue::Pool
Replace the ExecutionPools module with a Pool abstract class that the
thread and fiber pools inherit from, and that builds the right pool
for a worker, mirroring how Supervisor relates to ForkSupervisor and
AsyncSupervisor.
